+++ CHANGELOG.md
@@ -3,7 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+## XSTAMPP [3.1.1] - [Unreleased]
 
+### A-STPA Bugfixes:
+*  Fix Bug that when adding a new UCA in the Causal Factors Table it automatically adds a new Causal Factor which leads to save problems
 ## XSTAMPP [3.1.0] - 2018-09-12
 
 - Refactors documentation generation (README and CHANGELOG)